Stop exposing raw CRUD methods on Client object directly
The Client.Get(), Client.Delete() etc should not be public but instead
an implementation detail
